Start refactoring transcripts column into DefaultTranscriptsList
