+++ src/ensembl/src/content/app/entity-viewer/gene-view/components/default-transcripts-list/TranscriptsTable.scss
@@ -1,25 +1,15 @@
 @import 'src/styles/common';
+@import 'constants';
 
 .transcriptVisualisationColumnWrapper {
   position: relative;
-  margin-left: 228px; // FIXME: remove this when the table container is ready
-}
-
-.transcriptVisualisationColumn {
-  display: grid;
-  grid-auto-rows: 50px;
-  align-items: center;
-
-  svg rect {
-    fill: white;
-  }
 }
 
 .stripedBackground {
   position: absolute;
   top: 0;
   bottom: 0;
-  left: 0;
+  left: $left_column;
   right: 0;
 
   .stripe {
