+++ src/ensembl/src/content/app/entity-viewer/gene-view/components/default-transcripts-list/_constants.scss
@@ -1,0 +1,4 @@
+$left_column: 228px;
+$middle_column: 695px;
+$middle_column_right_gap: 27px;
+$right_column: 180px;
